pyproject.toml
[project]
name = "nonebot-plugin-maimai-qrhelper"
version = "0.1.0"
description = "nonebot-plugin-maimai-qrhelper"
readme = "README.md"
requires-python = ">=3.8, <4.0"
dependencies = [
    "nonebot-adapter-onebot>=2.4.3",
    "nonebot2[fastapi]>=2.2.1",
    "httpx>=0.27.0",
    "nonebot-plugin-alconna>=0.42.4",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_maimai_qrhelper"]
plugin_dirs = []
builtin_plugins = []

[tool.pdm]
